update packages and rename folders and files
