+++ AppFunction/LectureScreenFunction/deleteDuplicateLecture.js
