+++ AppFunction/LectureScreenFunction/readData.js
