+++ package.json
@@ -19,11 +19,11 @@
     "expo-constants": "~12.1.3",
     "expo-splash-screen": "~0.13.5",
     "expo-status-bar": "~1.1.0",
-    "expo-updates": "~0.10.13",
+    "expo-updates": "~0.10.15",
     "node-fetch": "^2.6.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "react-native": "0.64.2",
+    "react-native": "0.64.3",
     "react-native-dropdown-picker": "^5.1.28",
     "react-native-elements": "^3.4.2",
     "react-native-gesture-handler": "~1.10.2",
