+++ screens/lectureScreens/lectureApp.js
@@ -4,7 +4,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import { FontAwesome5 } from '@expo/vector-icons';
 
 // スクリーンの読み込み
-import SearchResult from './SearchResult';
+import SearchResult from './searchResult';
 import LectureScreen from './lectureScreen'
 import HomeLectureDetail from './homeLectureDetail';
 import EditLectureScreen from './editLectureScreen';
